--- packages/drawer/react.js
@@ -1,1 +1,0 @@
-module.exports = require('./dist/cjs/react')
